services/vanna/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
groq==0.4.1
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.27.0

